requirements.txt
celery==5.2.7
loguru==0.6.0
python-dotenv==0.21.0
redis==4.4.0
